Rewrote exception handling
Before, custom exceptions would throw an error themselves, because the error object would be handed over as a single string in some places. That has been resolved now, however: Custom errors can be triggered by using Error::exception(debug_backtrace(), 'error message');
This will provide a complete trace. It is not the most beautiful solution, but it works for now.